[mtk8167][clk] Implement all clock gates for 8167
The 8167 clock gate implementation was missing a number of
clocks. This CL fills in the remaining clocks while maintaining
existing functionality.

Tested by booting on cleo and observing that peripherals that
use clocks continue to operate.
Newly added clocks are yet untested but should work as their theory
of operation is identical to existing clocks.

Change-Id: I94387374e6ea00bf44f26d34d26a5abf13e532d7