[audio_core][e2e] Fix position calculation when aggressively downsampling
This CL addresses a longstanding bug in the calculation of source position
that would primarily have been exposed when downsampling by very high
ratios.

The fix to the product itself is surgical, and is specifically in the
classes outside the mixer (that "feed" it). Along with the fix, this CL
adds a set of audio fidelity tests (frequency response and SINAD) at the
3:1 down-sampling ratio.

This defect was exposed by the E2E tests, one of which originally spawned
a virtual audio output device that supported only 16 kHz.

Test: Manually verified the output and input mix paths, using
signal_generator and wav_recorder. Also added FR and SINAD test
cases for both Linear and Point samplers.

DNO-468 #comment

Change-Id: I5b07219ac7fda772c8aeca4e2f26ea59e0ce9818