[debugger] Suspend/Unsuspend thread for setting HW breakpoint.
This extends the suspend logic within a thread to support better
detecting on what state the thread was before the call. This is
important because currently suspended and blocked states are mutually
exclusive within zircon (ZX-3772).

This also extends the suspend call to allow blocking on the suspension
signal, which is needed for setting HW breakpoints.

Renamed the DebuggedThread::Pause call to the more apt
DebuggedThread::Suspend.

ZX-3772  # This requires checking for exception state.

TEST=Manual on x64 NUC.
     Both SW/HW breakpoints work as well as pausing/resuming.

Change-Id: Ieac518f47facf4a64b06c3536f886e65d01a8574