[zircon] Rename 'VMO clone' to 'VMO child'
In preparation for new types of VMOs for which 'clone' is not an
accurate description, rename 'clones' to 'children'. The renaming only
affects up to the syscall layer, as higher layers can still refer to
copy-on-write children as 'clones'.

This change adds a temporary vdso wrapper for the old zx_vmo_clone
syscall, which will be removed once the Go runtime is updated.

Change-Id: I35222e344d7fabdb64b58d1b59e12e2cc4a49ab9