[zircon][fvm] Port to Cpp style.
Update remnants from C to Cpp. This is mostly a style, not a functional
change.

This includes removing existing macros and replacing references to them
to constexpr.

Encapsulated logic for table entries (whether they are allocated or
something else) so we can remove a lot of implementation details from
the users (PSLICE_ALLOCATED, the flag mask, etc). This should make it
easier to change how we mark slices as allocated or free in the future.

Added unittests for entry types on fvm tables.

test= fvm-tests, CQ

Change-Id: I7b16dbe0faff9bbcf8e85b8fc75aa126572a08ea