[debugger] Thread exception handling reordering.
The flow now is easier to follow and reason about, as it leverages more
the exception decoding instead of doing twice.

Increased the priority of HW breakpoints on x64, as otherwise when
having both a SW and HW breakpoint on the same address could make the
debug agent to enter an exception loop.

TEST=Tested on both x64 NUC and arm64 VIM2:
     - Unit + Integration tests.
     - Manual cases: Stepping, SW + HW breakpoints (alone and together).

Change-Id: I9c3d2dbc025adda1552c06158606919937f7906b