[pkg_resolver] read static repo configs
This extends the package resolver to load static repository
config data into a RepositoryManager. When loading data off
of disk, it errs on being eventually consistent, where it will
ignore malformed configuration files in the config-data
directory, rather than erroring out the entire load.

Furthermore, it uses the following process to resolve conflicts
between repository definitions during the loading process. First,
all the config files are sorted in lexographical order, so the
load order is deterministic. Second, If there happens to be two
configs that are trying to define the same repository, it chooses
the one that comes from the filename that matches the repository
hostname. If both, or neither are in the same config file, then
the first config seen will be used.

Bug: PKG-612 #done
Change-Id: Ic3d100128407ebe29945a0e5aee2c74363c4257a