[fuchsia.net] Add fuchsia.net.namelookup FIDL interface
This adds an interface for doing hostname lookup, which is higer level
and easier to use than DNS, but provides enough features to implement
GetAddrInfo on top of.

The service that implements this interface will provide the features
that nsswitch and /etc/hosts provide on Linux (or at least, the subset
of those features that we choose to support). For instance, resolving
127.0.0.1 to localhost.

Change-Id: If455729d2ee290c11da3523ea87bb918ed5cbb84