[pseudo-fs] watchers: fix wire messages, tests
Turns out FIDL serialization is different from the current C++
serialization.  When I have added corresponding tests, FIDL
serialization was not able to deserialize a sequence of watcher events.

Appraently, FIDL deserialization can only support one top level object
per buffer, as it does not return the end point of deserialized data in
the buffer, even when not all the bytes have been used.  I have filed
FIDL-581 to fix that.

And I have also discovered, that the serialized form was not the same as
in Rust.

Change-Id: I28bb792c511c0aed826d501520bf7682942407db