[scenic] Make PaperRenderer2 the default.
This sets the default value of the 'scenic_use_new_renderer' build
arg to true.   This only affects the unshadowed case.

This has been tested by various members of the graphics and
performance teams (via the build flag).  This revealed a bug
in Flutter (FL-214) that manifested as significant Z-fighting.
The fix for FL-214 finally rolled into Fuchsia, and seems to
work properly.

We have no reason to suspect that the renderer itself behaves
improperly, so if any new graphical artifacts are observed,
the bug reports will ideally come with a dump of the Scenic
scene graph so that we can verify whether clients are at
fault.

SCN-670 #comment

Change-Id: I94f8535852ec53b4b205b9df726a49108b67b2ae