[debugger] Remove separator from Identifier.
Identifier needs to move to the symbol system so we can look things up
and return them in a reasonable format. This means that the dependencies
on the expression system (ExprNode) need to be removed.

This is part one which removes the separator tokens. These weren't
actually used for anything other than marking a separator was present so
this is a simplification also. There is now a separate flag for an
identifier in the global namespace.

Removes the comma after the separator in the Identifier debug format.
This wasn't helping anything and made the output more difficult to read.
This required updating a lot of test expectations.

Change-Id: I8ea5e8c047e2b45b349d8e4c0c12dca41098df93