[debugger] Move identifier name to a string.
All the strings in Identifier used to be ExprNode name tokens which make
it easy to correlate back to input. But we didn't really use this
facility and it's more useful to be able to use Identifier in the symbol
system that can't depend on the expression parser.

This simplifies Identifier and makes it easier to move.

Change-Id: Ie29c8b1c1eb36d84e47cf0f989e9902b3b0ade11