[usb] Add multi_tt to bus and hci ConfigureHub stubs.
Some USB implementations require the host controller to be programmed
with the knowledge of whether downstream USB hubs support single or
multiple transaction-translators.

This reads the requisite descriptors and passes the information to
ConfigureHub ultimately allowing the HCI driver to perform the
configuration.

This is currently a no-op for the xhci host driver, but will be used by
the mediatek host driver when I add downstream hub support.

Test: vim2 (xhci host)
Change-Id: I1bdc03ca6a78f5ec453ff3f449cf522c828d53cf