[cobalt] Initialize SysLogger
CB-285

This is in preparation for switching all uses of FXL Logger, which is
deprecated, for SysLogger. That switch is mostly not happening yet but
this CL allows us to make the switch incrementally. The new LOG_FIRST_N
functionality only exists in syslogger and we are going to use that
in Cobalt Core.

Change-Id: I95ee5cbba3c11bc04fb704d9729419e8b7d921cb