[zircon] Fix download-prebuilt script.
There is a bug in download-prebuilt script. This script compare
the output of "cipd ls fuchsia_internal" with "No matching packages."
to determine if user has internal access. However, When cipd is not
yet bootstrapped, buildtools/cipd will write "Bootstrap cipd" into
stdout. The compare logic in download-prebuilt script will fail.
This patch fix it by adding "tail -1", so only the last line of the
cipd output will be used when comparing with "No matching packages.".

Test: Local
Change-Id: I344af2616309095c493a3060720fda8dbef2039a