[fidl] [lint] Declarative tests
Added declarative test pattern so redundant code can be removed from
lint_findings_test.

Note during testing I discovered the caret pointing at the first
character of the token was off by one. Fixed in the CL as well.

I also added an option in TemplateString to strip variables that are
unmatched. So comparing with this option set vs. unset provides a way to
tell if any substitutions are missing. This was added to my test
validation method.

Test: fidl-compiler-test
Change-Id: If214a314d1d6859582ab27ae5ba627b3a05c7515