[host][fs][fvm] Way to dump usage details of storage entities
The patch introduces following three commands for host side blobfs, minfs
and fvm.
used-data-size - Gives number of blocks used by the data. This may include
                 space used by directory entries.
used-inodes    - Number of inodes allocated. This may be the sum of inodes
                 used for files and directories.
used-size      - This is sum of space used by data and sum of all
                 reservations for superblock, bitmaps, inodes and journal.
                 One way to interpret this is total-size minus used-size
                 gives current useable size.

Existing command `fvm size` retains only a part of it's original meaning
- the total number of bytes reserved for partitions.

To keep a check on growth of system image, it is recommended to track
all 4 of the above variables. Keeping a check on subset of these 4
variables may lead to "leaking" the growth information.

Change-Id: Ie9b675e5c36c07946255997443995dd81a4b8d4b