[minfs] Support CleanFlag unset/set on mount and unmount
This change implements the "clean" flag in superblock to help
determine if a filesystem was unmounted cleanly. We unset the value of
this flag on mounting with read-write options and on a clean unmount, we set this value again.

If during mounting, the value is not set, we simply print a warning for our
debugging purposes, but in future we can choose to run an intergrity checker
if need be.

Testing was done by dumping superblock fields on mount and unmount
and making sure the clean flag is set
and unset correctly.

Fixes: ZX-3891 #done

Change-Id: Idc1d57624058f2497f70051dd7ac88560efb19f0