[thinfs] fix nil pointer dereference.
This CL fixes a nil pointer dereference that happens in the
thinfs/zircon/rpc package.

The issue is that the `directoryWrapper is not correctly initialized,
not setting the cookies map. Upon calls to GetToken(), the dictionary is
dereferenced and that causes a panic.

The commit just makes sure that the cookies map is initialized
correctly, and adds a test for it.

ZX-3928: Done

TEST=/pkfs/packages/go_thinfs_tests/0/test/go_thinfs_test

Change-Id: I4686714f547f8a759d11dcc2e9414ce60e37707d