[fidlc] Fix to coding tables for xunion/table members
Certain types only need to be coded when they appear within envelopes,
e.g. primitives and arrays. We therefore need to know in which context
we generate coding tables to conditionally generate coding tables for
these types.

This worked for primitives because these are interned, such that
referencing these coding tables was sufficient. For arrays however, we
must generate the coding table, and then reference it.

Test: in fidlc and
Test: ./scripts/build-zircon-x64 && ./scripts/run-zircon-x64 -c zircon.autorun.boot=/boot/test/sys/fidl-coding-tables-test
Change-Id: I4c923cd2c6bc36c6e4e974b4a05cd139fded34d4