[kernel] Add shim from old zbi_cpu_config_t to SystemTopology.
When we handle the zbi_cpu_config_t type on ARM systems we will now
convert it to a new flat topology and feed it into the topology
subsystem. This allows us to have the Topology subsystem as the source
of truth for topology logic on all platforms. Eventually we will remove
the zbi_cpu_config_t alltogether.

ZX-3068
Test: Booted and "k ut all" on:
*Qemu arm
*Astro
*Vim2

Change-Id: Ia22eb29146b56e030d540dd92e469e48f0726e4b