[mdns] control trace compilation using gn arg
The mdns service has a feature that logs traffic. This feature is
controlled using mdns-util. The feature only works in the debug
version of the service to reduce the footprint of the release binary.

This CL adds a new gn arg (enable_mdns_trace) that determines whether
this feature is built in rather than release vs debug. Most people
build release, and it's a pain to switch the whole OS over to debug
just for the tracing feature.

TEST: build both ways
      $ fx shell mdns-util verbose

Change-Id: I4ba62bf54ce47916a7f7c9ef0ec1d2a8156eedbf