[sys] Annotate all calls of get_root_resource()
get_root_resource() is deprecated for all uses outside of platform bus
drivers, but a common development pattern is to search for code creating
physical vmos or other elements and use the same strategy in new code.
To avoid this I'm annotating every single call of get_root_resource() in
the system. These comments will be trivially removable at a later date.

Change-Id: I25bf0e970560d92efef681420d771f90925ba46a