[dev_finder] Issue mDNS queries on multiple ports (5353,5356 by default)
This teaches the dev_finder tool to issue mDNS queries on multiple
ports, responding to whichever port answers first. The -port flag now
accepts a comma separated list of port numbers. The default is to query
on port 5353 and 5356.  This way the dev_finder tool will keep working
while we (temporarily) shift the responder port around.

Change-Id: I7ce5c7a0898496f1c720b8f0130147317dadebb0