[media][drm] Clean up provisioning flow
Some DRM systems require the device to be provisioned specifically for
the provider. In the EME specification, DRM systems with such a
requirement are supposed to send an provisioning request the first time
any session requests a license. However, not all DRM systems implement
this automated flow, instead relying on a manual process of requesting a
provisioning request from the CDM. This CL eliminates the automated
provisioning flow in favor of a synchronous API to generate a request.

There are no consumers of this API yet, so no soft transition required.

Change-Id: Ie705a90d2442ffeccf03f1ac55a750b0642b6df5