[wlan-inspect] Change trait design for inspect log
The previous design works well for simple types but not for complex
types since the Inspect FIDL doesn't have a tree structure built-in.
Switch to a trait that allows writing to the node directly.

This commit also removes default implementations for u8 array types in
favor of having client explicitly choosing to log such types as a list
object or a bytes property using the provided wrapper types. In
addition, a couple WLAN types are implemented as well.

Test: fx run-test wlan-inspect-tests
Change-Id: I4ba553a940663043ec53f6d6bc1ebc0a7427a04d