[build] Fix integration with Zircon build when it uses Clang
The --strip-headers format (which is optimal for device storage use)
is incompatible with linking.  The Zircon build's `.so` files are in
this format when it builds with Clang-based variants rather than the
default "gcc" variant.  For C/C++ linking, the build system feeds
the right files (the `.so.debug` files) directly into the link step.
However, Rust and Go builds rely on -L... search paths and -l...
searches matching lib....so file names.  So this must be accomodated
in the directories set up by the build for -L... switches passed to
those compilers.

Bug: BLD-447 #comment Fixed --variant clang build
Test: fx set core.arm64 --variant clang && fx build
Change-Id: Ie7a030688b0a5f5fa19682a74238f7882189f865