[minfs] Restrict CleanFlag set/unset only on target
Currently Shutdown is only invoked on target.

See: system/ulib/fs/connection.cpp, "Connection::DirectoryAdminUnmount".
This is a privileged IPC request which causes the filesystem to shut down,
but it only runs on Fuchsia.

So this results in a somewhat asymmetric relationship with the clean bit:

Host + Target: Set clean on format.
Host + Target: Check clean on mount. Unset clean bit.
Target: Set clean bit on unmount.

This means that using the host-side tools,
doing any operation on the image (adding a file, listing contents, etc)
will make the image dirty.

In order to avoid this, we will be restricting the CleanFlag set/unset only to target

Fixes: ZX-3931 #done

Testing : Tested on target, as well as host to verify if we get the warning message.

Change-Id: If62acfb1d83acb99f7c334480d7d35ecb95fa1e0