[fdio] Fix unbinding top-level file from namespace.
Due to the ordering of vn lookups in fdio_ns_unbind, it is not possible
to unbind a directory at /* when more than one directory at that level
exists.

For example:
bind "/svc"
bind "/pkg"
bind "/test"
unbind "/test"

causes a debug assertion failure.

This change reorders the lookup of the next vnode in the chain to select
the correct node under the root for removal.

Change-Id: Icab5cf37229c93900606cbe2d8e64375c9b3818a