[debugger] Use new command line parser.
Uses the cmdline parser from Zircon. This is already used by zxdb and is
better suited to our needs than the fxl one.

This fixes a bug with the android unwind flag being broken. The
validated name didn't match the string that was read to enable.

Changes the unwinder flag to specify either "ng" or "aosp" which allows
us to change the defaults back and forth more easily.

Add the --unwind flag to the "fx debug" command.

Change-Id: I65783f125f4af81e1105373f4945a9745a270850