[debugger] Prefix search for function names.
Hooks up function name searching for autocomplete for locations ("break"
command, etc.). This also completes class and namespace names.

The behavior isn't perfect but this is a good first step we can iterate
from.

Adds some additional test infrastructure to be able to mock targets, and
generalize the existing process mocks.

Adds a name to the FoundName class since currently namespaces and
templates have no way to return the matched name when there is more than
one. Previously we were only doing exact matches for these kinds so a
separate name was not needed.

Change-Id: I44e88165ff66a7b9a6766ff5979da286af6d1bbf