[magma] Add a libmagma_complete library
With a normal static library gn doesn't link any object files into it,
but just passes them on to shared libraries that include this file. That
doesn't work if we just to use the static library in a build system
that's not gn.

We can add a new static library with complete_static_lib set that's
usable outside the build system. We can't just make the existing library
complete, because that causes ODR violations in current code.

Change-Id: I076f3f7a0e5aed1f27e276aefef9aa735a01c82d