[syscalls][banjo] Replace syscalls.abigen with syscalls.banjo
This is intended as a transitional step on replacing abigen with banjo.
The goal of this backend is for it to be a relatively short-lived as a
migration path:

1. Create abigen backend for banjo
2. Replace hand-written abigen with output of equivalent .banjo file as
   a Zircon build step
3. Create additional backend(s) that generate the current outputs of
   abigen, but as there are currently 13, proceed one or a handful at a
   time
4. Once the build is consuming files generated directly from
   banjo, remove the abigen backend (and the abigen tool).

This CL is "Step 2".

This is intended to a be no-op functionality-wise. One notable change is
that syscalls.abigen is no longer included in the SDK, however it is
believed that this is OK and we are not aware of any consumers.

ZX-3882 #comment [syscalls][banjo] Replace syscalls.abigen with syscalls.banjo

Test: build twice once before this CL and once after, and confirm that
all 13 (!) outputs of abigen are identical.

Change-Id: I2517b87e0c9fe88266aaeb3a485f3042b1131eae