[debugger] Implement "pause" feedback.
Prints the stopped thread/process status after doing a "pause" command.
Previously there was no feedback which made the command ambiguous about
what actually happened. Some logit attempts to print the current thread
backtrace as long as the current thread was among those paused.

To support this, pausing is now synchronous and the updated thread info
for the affected threads is passed in the reply.

To support sending information for many process' threads in one vector,
the process ID is added to the ThreadRecord. This simplifies some other
places as nearly every place a ThreadRecord was used required a process
KOID to be passed separately. This required updating many files but the
changes are mechanical.

Filling a ThreadRecord is now moved to the DebuggedThread class instead
of having a standalone function that takes many arguments. It meant the
unit test for this function be updated significantly.

Change-Id: Iaf1faeee7b081207f7d095ee5c941adad328106d