[dev][xhci] Return error if command queue is full
XHCI now checks for the command queue being full and will return an error
if that happens.
There was some suspicion that the command queue was overflowing due to
errors in USB ethernet, but closer examination shows that this did not
actually happen. But now if it ever does, we will handle that error case
more gracefully.

TEST: manual testing on astro
Change-Id: I424aaf6464ef791133b526f70037cf034b5de9ba