[handle][bti][pmt] Create() KernelHandles
This one was a bit trickier than normal due to multi-part setup during
Create() as well as making sure teardown doesn't try to recursively grab
the BTI lock on initialization failure.

I think overall it's still a win though, particularly because
UnmapFromIommuLocked() is now called exactly once in on_zero_handles()
instead of potentially being called from several places in the code.

Also I think there was a race condition in the destructor that this
fixes. The comments indicated it was safe to check for list membership
without a lock, but if the following item in the list is removing itself
concurrently, it could be modifying our next_ pointer at the same time
that we're trying to read it.

ZX-3612

Test: runtests
Test: run-zircon-x64 -b -k -z build-gcc/core-tests-x64.zbi
Test: break Create() in a few places and make sure it works as expected

Change-Id: I98948ef40f925ac2ddfe9928dc4df5c321741daf