[virtcon] Make virtcon host a fidl service for sessions.
the virtcon (virtual-console) binary used to provide new sessions
through a convoluted path through device coordinator. This will make it
expose a fidl service directly for clients to request a new session.

This removes device coordinator from the operational loop and will make
it easy for the virtual-console binary to stand on it's own once we have
something else to start it.

It is a little messy due to the virtcon binary using the "port" library
heavily and fidl being built around the async_loop.

Test:
 * Booted on vim2, core.arm64. Virtual consoles work as expected.
 * Booted astro, looks normal, "k ut all" and "runtests -a" pass.

ZX-3403 # Get virtcon out of the way.

Change-Id: I09e66a7c5bc91fb41b6bfb99510b5be1aa609d57