[perfcompare] Apply Bessel's correction when calculating standard deviation
Our use of z-test confidence intervals here is not very good, but this
makes it a bit less bad.  This will widen the confidence intervals a
little, especially when using smaller sample sizes -- and we recently
switched to treating the sample size to be 30 rather than 1000.

Also add some sanity checks for error cases.

Bug: IN-646
Test: "python perfcompare_test.py"
Change-Id: I46aa0ce4069bb2afc8d26653b6b9c3bdc02a525b