[debugger] Thread specific SW breakpoints.
Though the UI was implemented, the debug agent could not handle a
breakpoint set for only a specific thread. This implements the logic
that a thread queries the breakpoint if it should apply to it and resume
if that's not the case.

Moved the SW breakpoint logic to its own file as process_breakpoint.cc
was getting hard to manage by itself, plus this is consistent with
hardware_breakpoint.

TEST=Manual, unit and integration.
     Future CL will have an integration test for this case.

Change-Id: I5f09b1f47cb19ce3316bfb31ee63f5312c0970d8