[bt][a2dp-sink] Detect socket close on AVDTP close
When a peer sent us a close signal, a 3 second timer would wait for the
peer to close the transport (as in AVDTP spec section 6.3) before
checking.

Instead complete the operation when the peer disconnects the transport
socket.

This mitigates a race condition where the a2dp-sink could mistakenly
add a second RemotePeer tracking the same peer while in the wait, and
that new peer would be removed, causing a panic when the peer couldn't
be found when it sent commands.

Test: connect to bt-a2dp-sink from android, reconnect quickly
BT-730 #done

Change-Id: Ia4adeced5e49c225228bbd0ba5314021203e5034