[pm] newrepo produces a fully ready repository
newrepo was only preparing staged repository data, which meant that after a
newrepo invocation the repository was not in a "servable" state. Newrepo now
commits the empty target set, to ensure that a newly formed repository is
entirely servable.

This change also updates the flags somewhat to unify between `pm newrepo` and
`pm publish`, but retains support for the old flags. In the case where a user
supplies both the `-repo` and the `-r` flag to publish, the last flag in argv
wins.

Bug: DX-972 #done
Change-Id: Ib15ae26d5a5cb3480e24d8e7fecdd2966189b9cd