[zircon] Rename 'VMO clone' to 'VMO child'
In preparation for new types of VMOs for which 'clone' is not an
accurate description, rename 'clones' to 'children'. The renaming only
affects up to the syscall layer, as higher layers can still refer to
copy-on-write children as 'clones'.

This change adds a vdso shim for the old zx_vmo_clone syscall and leaves
the old flags/zx wrapper. These will be removed once clients are
migrated to the updated API.

Change-Id: I8319d8ab379b7e57905ec3f6ece86fd6e0d401d1