[text] Switch TextFieldState from struct to table
Title says it all. We'd like this to be a table so that further upgrades
to this state object are easier.

This change also adds an alternative version of TextFieldState to three
Rust libraries, where the required fields on TextFieldState (document,
selection, revision) are T instead of Option<T>. Into/TryInto is also
implemented back and forth from the two types.

You'll notice that this implementation is duplicated in three
text_field_state.rs files, one for each project. I thought perhaps this
was better, since if we make changes to the table, each of these three
projects may want to upgrade to the new implementation separately?
Whereas if they all pulled text_field_state.rs from some common
location, we'd have to upgrade them all at once.

Change-Id: Icbb67498615b35e39825278a5c25c3d62fe7b0a0