[wayland][core] Use String::from_utf8_lossy.
When opening an image file, Firefox will set the window title to a
string that inculdes the image dimensions using what appears to be the
unicode multiplicaiton sign (code point 0x00d7). The specific encoding
supplied by Firefox does not appear to be valid UTF8.

Work around this by not panicing and simply truncating strings that are
not valid UTF8. PD-204 tracks finding a more proper solution here that
will allow us to move back to more strict UTF8 parsing.

Test: Opened image file in Firefox and observed no panic in the bridge.
Change-Id: I22f4f81d1865e98a9dfc317536fb3e9ab6199491