[benchmarking] Invert input/trace foreground/background-ness
Run input as a background job and trace record as a foreground job.
This allows us to block on trace record, which ensures that the trace
file is written by the time we start trying to process it.
Unfortunately there's still a race between input and trace record
startup, however I strongly prefer that race over what we had before.

Test: CQ

Change-Id: Id016219e305ae662ec8ed45681ca057131da8187