[fidlc] Improved support for anonymous names
Since we have anonymous names which do not come from source files, e.g.
the name of the built-in uint32 type, we must handle absence of location
all the way down to how we report errors.

One benefit is the removal of the is_anonymous() method on Name:
callers should always assume that names could be anonymous, and this
should be hidden by the Name object model.

Change-Id: Ib47d23c93d893c3fefc6093d9162264ec492e3ef