[debugger] Test that thread-specific breakpoint works.
This integration tests spawns a number of threads and verifies that only
the correct thread hits the breakpoint and all other continue untouched.

The actual logic was implemented in a previous CL.

Added a lot of debug statements in order to track down problems. As
usual, these are deactivated in normal running of the agent.

TEST=Integration :) (fx run-test debug_agent_tests)

Change-Id: If2561a9a7bf2bb60702023b808950e87c92f5894