[paver] Rework paver lib interface.
The paver lib was to be closer to the interface it will be once paving
is exposed as a fidl service. The next change will hook it up to the
paver service, and the last will switch the install-disk-image binary to
use the paver service, instead of of directly invoking the library
functionality.

ZX-3861

Tested: Paved NUC, astro, vim2, and pixelbook

Change-Id: Ic705e76af70be1153998888e93250dbf8be21fda