[debugger][x64] Added HW watchpoint test.
This creates a simple application to test that the low level thread
state API can be used to successfully trigger a HW watchpoint. This
binary is able to trigger a HW exception.

This binary is made for all archs though only the x64 side is implemented.
Ported the HW breakpoint binary to the same layout, having only the
arm64 side implemented.

TEST=Manual on NUC (x64).

Change-Id: Id201e9024626e213da1d4d5f3697d854e4fcd417