[zircon][system] Extract scsilib from virtio into a common library
scsilib is a library to construct a Fuchsia block device on a SCSI
Direct Access Block Device. It implements some of the complex SCSI-3
probe sequence and is used by virtio-scsi.

For expediency it was originally written as part of virtio/scsi;
extract it to a common location so UMS and UASP and UFS could use it.

ZX-2314

Tested:
./scripts/run-zircon-x64 -- \
    -drive if=none,id=hd,file=/tmp/blk.img \
    -device virtio-scsi-pci,id=scsi  \
    -device scsi-hd,drive=hd,scsi-id=1,lun=1 \
    \
    -drive if=none,id=hd2,file=/tmp/blk2.img \
    -device scsi-hd,drive=hd2,scsi-id=2,lun=1

$ lsblk
ID  SIZE TYPE             LABEL                FLAGS  DEVICE
000  16M                                              /dev/sys/pci/00:02.0/virtio-scsi/scsi-disk-1-1/block
001  16M                                              /dev/sys/pci/00:02.0/virtio-scsi/scsi-disk-2-1/block
$

Change-Id: Id6480a49ddc870fccc03e6a6311772901fbeee2e