[pseudo-fs] Rename "{fuchsia-vfs-,}pseudo-fs"
Repeating "fuchsia-vfs" twice in the path is redundant.  When I have
created the package I was not sure about the naming rules of the
generated artifacts and just used other packages as exmaples.

Now I know how to change BUILD.gn in such a way that the generated
artifact is still named fuchsia_vfs_pseudo_fs_lib, but the folder path
duplication is avoided.

Ran test for component_manager and pseudo-fs.

Change-Id: I38e15b9e7540c47b9d2893fa5cfd7685e15872b1