[debugger] Make sys-info work for minidump
Minidump has most of the interesting info here. The only thing we don't
have is the memory size. We update the UI so we can output unknown in
that case. We don't have information on breakpoints or watchpoints
either, but arguably "zero" is the correct answer in either case for a
core dump.

Change-Id: Ib7d48e86b96bddd6fb0ae3511c83f5ecd10f7e31