[fs] Check path component length in vfs
Previously, the vfs validated the entire path length but did not
validate the length of each individual components. When a client opens a
path with a component larger than NAME_MAX, it would trigger debug
assertions in minfs, which expect each path component to be within that
limit.

ZX-3924 #done

TEST: /boot/test/fs/fs-test

Change-Id: I3d64f2c253591eab5dff9a8fd4b1d7abefbf6388