[zircon][scsilib] Unit test of scsi::Disk
scsi::Disk in scsilib connects a SCSI Direct Access Block Device to
the Fuchsia block device layer, translating queued Ops to SCSI commands.

Write a unit test using an in-test fake controller. Initially just
test that a controller can be created and that INQUIRY and READ CAPACITY
(16) are executed during startup. Also test that the DDK block size is
the same as the size reported via the SCSI commands.

Subsequent test cases will check the properties of the block device and
execute I/O.

ZX-2314
ZX-3757

Change-Id: I3625a92a8ffa302b5b9e5667aaeed30a2ef5c0d9